Move AddLesson into AddLessons--no need for separate components anymore
