+++ client/src/components/index.js
@@ -1,4 +1,3 @@
-import AddLesson from './AddLesson';
 import Carousel from './Carousel';
 import Categories from './Categories';
 import CheckoutForm from './CheckoutForm';
@@ -7,7 +6,6 @@ import Navbar from './Navbar';
 import Recommended from './Recommended';
 
 export { 
-  AddLesson, 
   Carousel, 
   Categories, 
   CheckoutForm,
